# PERF/REGR: isin slowdown for masked type